Moving to unsigned int parse to match c#

+++ cv-data-service-library/src/test/java/com/trihydro/library/service/SdwServiceTest.java
@@ -116,7 +116,7 @@ public class SdwServiceTest {
 
     @Test
     public void deleteSdxDataBySatRecordId_nullRecordIds() {
-        HashMap<Long, Boolean> results = SdwService.deleteSdxDataBySatRecordId(null);
+        HashMap<Integer, Boolean> results = SdwService.deleteSdxDataBySatRecordId(null);
         PowerMockito.verifyStatic();
         Utility.logWithDate("Attempting to delete satellite records failed due to no satRecordIds passed in");
         assertNull(results);
@@ -124,7 +124,7 @@ public class SdwServiceTest {
 
     @Test
     public void deleteSdxDataBySatRecordId_emptyRecordIds() {
-        HashMap<Long, Boolean> results = SdwService.deleteSdxDataBySatRecordId(new ArrayList<String>());
+        HashMap<Integer, Boolean> results = SdwService.deleteSdxDataBySatRecordId(new ArrayList<String>());
         PowerMockito.verifyStatic();
         Utility.logWithDate("Attempting to delete satellite records failed due to no satRecordIds passed in");
         assertNull(results);
@@ -135,7 +135,7 @@ public class SdwServiceTest {
         List<String> satNames = new ArrayList<String>();
         satNames.add("A9184436");
         Mockito.when(mockConfig.getSdwApiKey()).thenReturn(null);
-        HashMap<Long, Boolean> results = SdwService.deleteSdxDataBySatRecordId(satNames);
+        HashMap<Integer, Boolean> results = SdwService.deleteSdxDataBySatRecordId(satNames);
         PowerMockito.verifyStatic();
         Utility.logWithDate("Attempting to delete satellite records failed due to null apiKey");
         assertNull(results);
@@ -147,7 +147,7 @@ public class SdwServiceTest {
         satNames.add("A9184436");
         Mockito.when(Utility.getSdxUrlConnection(isA(String.class), isA(URL.class), isA(String.class)))
                 .thenThrow(new IOException());
-        HashMap<Long, Boolean> results = SdwService.deleteSdxDataBySatRecordId(satNames);
+        HashMap<Integer, Boolean> results = SdwService.deleteSdxDataBySatRecordId(satNames);
         assertNull(results);
     }
 
@@ -169,7 +169,7 @@ public class SdwServiceTest {
         Mockito.when(mockUrlConn.getResponseCode()).thenReturn(200);
         Mockito.when(Utility.getSdxUrlConnection(isA(String.class), isA(URL.class), isA(String.class)))
                 .thenReturn(mockUrlConn);
-        HashMap<Long, Boolean> results = SdwService.deleteSdxDataBySatRecordId(satNames);
+        HashMap<Integer, Boolean> results = SdwService.deleteSdxDataBySatRecordId(satNames);
         assertNotNull(results);
         assertThat(results, IsMapContaining.hasKey(-1));
     }
